Xcode project file
 Committer: Alexandru Naiman <alexandru.naiman@icloud.com>

+++ src/Screen.cpp
@@ -11,7 +11,7 @@ void Screen::Init ()
 	std::size_t width = Window::GetWidth ();
 	std::size_t height = Window::GetHeight ();
 
-	_surface = SDL_CreateRGBSurface(0, width, height, 32,
+	_surface = SDL_CreateRGBSurface(0, (int)width, (int)height, 32,
 		0x00FF0000, 0x0000FF00, 0x000000FF, 0xFF000000);
 }
